requirements.txt
Django
djangorestframework
django-cors-headers
redis
celery
psycopg2-binary
requests
django-celery-beat
uWSGI
